requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
gTTS==2.3.2
python-dotenv==1.0.0
SpeechRecognition==3.10.0 